setup interface step 1
